organised dir structure and did minor refactoring
