+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 //import ComingSoon from "react-coming-soon"
-import App from './App';
+import App from './containers/App';
 import 'tachyons';
 import * as serviceWorker from './serviceWorker';
